remove ng1 components from app
