+++ generators/client-2/templates/src/main/webapp/app/_app.main.ts
@@ -1,7 +1,7 @@
 import { UpgradeAdapterRef } from '@angular/upgrade';
 import { upgradeAdapter } from './upgrade_adapter';
 import { UrlRouter } from 'ui-router-ng2';
-import './app.module';
+import './app.ng1module';
 
 function readyFn(ref: UpgradeAdapterRef) {
     let $urlRouter = ref.ng2Injector.get(UrlRouter);
