+++ generators/client-2/templates/src/main/webapp/app/_upgrade_adapter.ts
@@ -2,7 +2,7 @@ import * as angular from 'angular';
 
 import { UpgradeAdapter } from '@angular/upgrade';
 import { forwardRef } from '@angular/core';
-import { <%=angular2AppName%>AppModule } from './app.ng2module';
+import { <%=angular2AppName%>AppModule } from './app.module';
 
 export var upgradeAdapter: UpgradeAdapter = new UpgradeAdapter(forwardRef(() => <%=angular2AppName%>AppModule));
 upgradeAdapter.upgradeNg1Provider('$stateParams');
